Fix relocation problems
many programs ship with *-config tools or *.pc files to allow automatic
configuration. Most of the *-config tools do the correct thing and
replace the prefix with the correct one not all do so we fix them with
sed.

*.pc files for pkg-config are more tricky since they are not evaluated
by a shell. So we now have our own pkg-config wrapper which a) adds the
correct directories to the lookup path and b) defines a variable
BELLE2_EXTERNALS_DIR which can be used in .pc files. Also, .pc files are
fixed to use this variable.

Also, we fix a few .cmake files which aren't used but just to be save.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/externals@21297 b38c6c8f-bee8-465d-8448-9abe6915d496