Avoid problems with command line goals
As we now have common and <opt> compilation people could call "make gcc"
and make would compile it in opt as the command line goal overrides the
implicit goal selection. I tried a few options to get rid of these
problems but nothing worked to my satisfaction. What we could do is
redefine the BELLE2_EXTERNALS_OPTION in the makefile according to the
package but that seems hacky.

Or we could just split into two Makefiles, one for the common stuff and
one for the rest, and maybe a third with the variables needed by both.

For now, calling make gcc (or any other target which is compiled in
common mode) will raise an error and tell the user what to do. If
someone is so clever to call make Linux_x86_64/opt/bin/gcc directly well
then so be it, they either know what they are doing or don't deserve it
any better.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/externals@21300 b38c6c8f-bee8-465d-8448-9abe6915d496