Change download script
The script now takes two mandatory arguments: directory name and archive
name and additional a list of urls. It will download the archive in
src/ and verify it's sha256 checksum before extraction [1].

It will extract in a temporary directory and rename the extracted
directory to the src/<directory name> automatically.

This shortens the repetitive cd && download && mv cycle we had
everywhere.

[1] Requested by Christian Pulvermacher, and it makes sense to not just
downloaded stuff from the internet and trust it. When packages get
updated the checksum can be ignored by setting
BELLE2_EXTERNALS_IGNORE_CHECKSUM and running "make src && make checksum"

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/externals@21315 b38c6c8f-bee8-465d-8448-9abe6915d496