Fix GCC search path
GCC will *always* look into ${prefix}/include *after*
/usr/local/include, even if the directory is specified with -I. This
means that /usr/local/include was searched before the Belle2 externals
include dir.

The only solutions I found are:
  a) patch gcc to ignore ${prefix}/include
  b) specify a non existing path instead of /usr/local/include
  c) compile with sysroot option which needs a copy of all system
     includes
  d) install gcc in a different directory
  e) set the prefix to something different but adjust all other paths
     (exec-prefix, datarootdir, cxx-include-dir) to the usual ones.

a) and b) are very hacky and probably break something somewhere, c) is a
lot of effort to get the system includes correctly and I'm not even sure
it will solve the problem adequatly. It's mainly used for
cross-compilers. d) is possible but introduces yet another directory we
need in the path.

e) is easy enough to do and keeps the file layout basically as it was.
However, all directories need to be inside of prefix so the only
sensible option we have for the prefix is $(BELLE2_EXTERNALS_DIR)/$(BELLE2_ARCH).

Still, it seems to work so let's go with e). GCC creates an empty
include dir in the prefix directory so we make sure it's empty and
delete it.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/externals@21759 b38c6c8f-bee8-465d-8448-9abe6915d496