Remove libboost_python3
it's identical to libboost_python but causes problems when both are
loaded at the same time. So let's remove one of them to make sure this
doesn't happen.

To make sure that we don't pick up a libboost_python3 from the system we
create symlinks to the libboost_python library

closes #2027

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/externals@28213 b38c6c8f-bee8-465d-8448-9abe6915d496