Setup Environment variables needed by git
Apparently git is one of the few non-relocatable programs which
hardcodes absolute paths. But it seems there are environment variables
for all necessary directories so set them correctly.

Note: There is a RUNTIME_PREFIX define in the Makefile but it's intended
for windows so I'll have to see if it works on Unix.