Make cppcheck relocatable
cppcheck compiled the absolute path to the configuration files into the
binaries. This was not noticed before as cppcheck is flagged as optional
and thus I didn't test relocatability (people usually don't move the
externals they compile themselves).

This commit adds a patch to cppcheck to get the configuration file path
from the $BELLE2_EXTERNALS_DIR environment variable.