Update ROOT to 6.06/08
ROOT now depends on lzma but brings it's own version so
-Dbuiltin_lzma=ON was added.

Rave had problems compiling with the new version due to the extremly
ugly header removal and linking we do. It's now fixed by even uglier
linking of the root includes to rave and overriding the LDFLAGS
environment variable to add the root libraries to rave.

Fixes now included upstream release:
 - ROOT-7872 PyROOT: wrong return value when calling a c++ function
   returning short (patches/root-pyroot-short.patch)
 - ROOT-7940 Incorrect handling of LLVM Threading
   (patches/root-disable-threaded-llvm.patch)
 - ROOT-7912 Failure with -Dbuiltin_davix=ON on 64bit Ubuntu
   (patches/root-davixcompile*.patch)
 - ROOT-7903 ROOT 6.06.00 does not compile on x86 32bit
   (patches/root-minmax-engine-32bit.patch)
 - ROOT-8094 Remove C++1z syntax/GNU ext from math/mathcore/src/mixmax.h
   (patches/root-mixmax-nohexfloat.patch)

Still open:
 - ROOT-7640 Problem importing TH1D in ROOT 6 with python 3.5
   (patches/root-python3.patch)
 - ROOT-7638 Iteration not working properly in python3
   (patches/root-python3.patch, closed upstream but not included in 6.06
   patches)