Fix python-packages target
`pip install --upgrade` causes dependencies to be upgraded as well. So we
either need to specify setuptools and pip as well or install without
upgrade.  At the moment that seems the better option.

Also python-packages is now an empty target just depending on the
ipython executable to only run it once like most other packages.