Some python packages are not available as source package
So downloading source packages fails and pip3 stops processing. While
this should not be a big issue as wheels are preferred for installing
anyway this breaks the intention to download both source and wheel to be
on the safe side.

So now we download the packages one by one and ignore errors :/