Add root-numpy and root-pandas
This also requires to have two steps of pip install:

1. the usual one right after compiling python
2. a second one after compiling ROOT (root wants/needs numpy)

The second step needs to install into the opt/debug folder, otherwise
all the include paths/ld library paths are not setup correctly to find
ROOT so we get a second site directory in opt or debug

So let's add a sitecustomize.py which will add the corresponding site to
the path automatically.

I added a second step to the sitecustomize which will also add an
existing `python-packages` folder in either the central release or the
local release directory to the path as well. That way one could install
additional python packages with
`--target="$BELLE2_LOCAL_DIR/python-packages"`.